+++ libcovebods/lib/common_checks.py
@@ -36,13 +36,17 @@ class RunAdditionalChecks:
     def __init__(self, json_data):
         self.json_data = json_data
         self.person_statements_seen = []
+        self.person_statements_seen_in_ownership_or_control_statement = []
         self.entity_statements_seen = []
+        self.entity_statements_seen_in_ownership_or_control_statement = []
         self.output = []
         self.possible_out_of_order_statements = []
 
     def run(self):
         self.person_statements_seen = []
+        self.person_statements_seen_in_ownership_or_control_statement = []
         self.entity_statements_seen = []
+        self.entity_statements_seen_in_ownership_or_control_statement = []
         self.output = []
         self.possible_out_of_order_statements = []
 
@@ -69,7 +73,11 @@ class RunAdditionalChecks:
             # TODO here check statement_id is a valid Id
 
             if statement_id:
-                if statement_type == 'ownershipOrControlStatement':
+                if statement_type == 'entityStatement':
+                    self._check_entity_statement_second_pass(statement)
+                elif statement_type == 'personStatement':
+                    self._check_person_statement_second_pass(statement)
+                elif statement_type == 'ownershipOrControlStatement':
                     self._check_ownership_or_control_statement_second_pass(statement)
 
         # Turn checks into output
@@ -95,6 +103,7 @@ class RunAdditionalChecks:
             interested_party_described_by_entity_statement = interested_party.get('describedByEntityStatement')
             interested_party_described_by_person_statement = interested_party.get('describedByPersonStatement')
             if interested_party_described_by_entity_statement:
+                self.entity_statements_seen_in_ownership_or_control_statement.append(interested_party_described_by_entity_statement) # noqa
                 if interested_party_described_by_entity_statement not in self.entity_statements_seen:
                     self.possible_out_of_order_statements.append({
                         'type': 'entity_statement_out_of_order',
@@ -103,6 +112,7 @@ class RunAdditionalChecks:
                         'seen_in_ownership_or_control_statement': statement.get('statementID'),
                     })
             if interested_party_described_by_person_statement:
+                self.person_statements_seen_in_ownership_or_control_statement.append(interested_party_described_by_person_statement) # noqa
                 if interested_party_described_by_person_statement not in self.person_statements_seen:
                     self.possible_out_of_order_statements.append({
                         'type': 'person_statement_out_of_order',
@@ -114,6 +124,7 @@ class RunAdditionalChecks:
         if isinstance(subject, dict):
             subject_described_by_entity_statement = subject.get('describedByEntityStatement')
             if subject_described_by_entity_statement:
+                self.entity_statements_seen_in_ownership_or_control_statement.append(subject_described_by_entity_statement) # noqa
                 if subject_described_by_entity_statement not in self.entity_statements_seen:
                     self.possible_out_of_order_statements.append({
                         'type': 'entity_statement_out_of_order',
@@ -121,6 +132,20 @@ class RunAdditionalChecks:
                         'entity_statement_out_of_order': subject_described_by_entity_statement,
                         'seen_in_ownership_or_control_statement': statement.get('statementID'),
                     })
+
+    def _check_entity_statement_second_pass(self, statement):
+        if statement.get('statementID') not in self.entity_statements_seen_in_ownership_or_control_statement:
+            self.output.append({
+                'type': 'entity_statement_not_used_in_ownership_or_control_statement',
+                'entity_statement': statement.get('statementID'),
+            })
+
+    def _check_person_statement_second_pass(self, statement):
+        if statement.get('statementID') not in self.person_statements_seen_in_ownership_or_control_statement:
+            self.output.append({
+                'type': 'person_statement_not_used_in_ownership_or_control_statement',
+                'person_statement': statement.get('statementID'),
+            })
 
     def _check_ownership_or_control_statement_second_pass(self, statement):
         interested_party = statement.get('interestedParty')